+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import java.time.LocalDate;
 import java.util.List;
