+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.dao;
 
 import java.time.LocalTime;
 import java.util.List;
